Updated computer player digging algorithm.
Modified the algorithm so that if it reaches a slab connected
to dungeon, it sets this place as starting point.

This should prevent computer player from damaging walls within
the dungeon while digging.

+++ src/globals.h
@@ -196,7 +196,7 @@ typedef short CrStateRet;
 /** Type which stores CrCkRet_* values. */
 typedef short CrCheckRet;
 /** Type which stores Job_* values. */
-typedef unsigned long CreatureJob;
+typedef unsigned long long CreatureJob;
 /** Creature instance index, stores CrInst_* values. */
 typedef short CrInstance;
 /** Creature attack type, stores AttckT_* values. */
